Create a Ui interface
